优化：用户不能登录后台，管理员不能登录前台

+++ app/Http/Requests/StoreWareRequest.php
@@ -33,7 +33,7 @@ class StoreWareRequest extends FormRequest
     {
         return [
             'product_id' => ['required', 'exists:products,id'],
-            'quantity' => ['required', 'int', 'max:999', 'min:1'],
+            'quantity' => ['required', 'numeric', 'max:999', 'min:1'],
             'type' => ['string'],
         ];
     }
